[prpl][feed-mediatek][WiFi7][Fix critical patch code defect issue]
[Description]
Fix critical patch code defect issue

When there is no STA connected, the RSSI signal of a frame calculated
by mt76_rx_signal becomes 0, as the RSSI of the 5th RX is set to 0.
Therefore, set the RSSI of the 5th RX to the mininum value = -128 to avoid
polluting the RSSI signal.
For example, the auth and assoc resp frame sent from the STA will have
an RSSI of 0, which causes the rssi_reject_assoc_rssi
check to fail due to inaccurate rssi values.

Kernel6.6/MT76:
v2: 0108-cp-mtk-mt76-mt7996-add-rssi-dump-of-5th-rx-support.patch

[Release-log]
N/A

Change-Id: Ic80bda745f2421937be30d610da6f1216db30e74